@@ -1,5 +1,5 @@
-using Altinn.Authorization.ABAC.Xacml;
 using System.Threading.Tasks;
+using Altinn.Authorization.ABAC.Xacml;
 
 namespace Altinn.Authorization.ABAC.Interface
 {
