@@ -1,9 +1,9 @@
-using Altinn.Authorization.ABAC.Constants;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
 using System.Text;
+using Altinn.Authorization.ABAC.Constants;
 
 namespace Altinn.Authorization.ABAC.Xacml
 {
